% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reader_Function.R
\docType{data}
\name{list.of.packages}
\alias{list.of.packages}
\title{A Data read-in Function}
\format{An object of class \code{character} of length 4.}
\usage{
list.of.packages
}
\arguments{
\item{pth}{The path of the parent folder for the nested data files to be read from}
}
\description{
This function allows you to read all the data files that reside within a nested folder structure.
}
\examples{
fl_read()
filename_resolution()

Description:
Provide the starting folder path and the system will read all relevant files within this nested folder.
1.	Essentially, this package was designed to read files of the format (csv/xlsx/xls), so any other files that reside within the folder structure, eg. PDF, txt, etc; will be ignored.
2.	The naming convention of the read files will be:
  a.	For CSV:	 ProjData_<FolderName1>_<FolderName2>_<FolderName3>._filename
  b.	For XLSX/XLS:	ProjData_<FolderName1>_<FolderName2>_<FolderName3>._filename_SheetName1
ProjData_<FolderName1>_<FolderName2>_<FolderName3>._filename_SheetName2
3.	A report of files which could not be read (maybe due to corrupt data) will be generated for reference in the end.
}
\keyword{csv,}
\keyword{xls}
\keyword{xlsx,}
